redirect old trigger-service page
It still exist (clenup is not perfect on the docs bucket) and is wrong.
Another option would be to manually delete the file on s3, but since
people seem to still find it somehow, I think a redirect to the correct
page is better than a 404.

CHANGELOG_BEGIN
CHANGELOG_END