Change `CODEOWNERS` entries for @stefanobaghino-da's (#13812)
changelog_begin
changelog_end

In my new role, my main concern is providing context and get out of the
way of far more skilled and technically-involved engineers. As such, for
the time being I decided to reduce my subscriptions to the bare minimum.

That boils down to keeping an eye on `//docs`. This doesn't mean I don't
want to be involved in code anymore. Please do mention me in PRs if you
believe my attention is required.