remove workspace_status.sh
This script was introduced to control the date printed in the sitemap.
The sitemap is now produced in the docs repo, so there's little point in
keeping this mechanism.

Moreover, the stamp on the docs build means it gets rebuilt on every
change to the date, i.e. every day, which isn't good for caching.

CHANGELOG_BEGIN
CHANGELOG_END