split releases: add parameter
This adds parameters to the "split release" job. At the moment, when
run, the job just builds the current commit with a snapshot version.
With this change, we'll be able to build a split release for any commit
with any version tag.

The reason for this change is we want to trigger this job from the
assembly repo, to reduce the number of manual interventions needed in
the release process.

CHANGELOG_BEGIN
CHANGELOG_END