start: work with console
This fixes the bug reported by `kctam` on [the forum]: that

[the forum]: https://discuss.daml.com/t/daml-start-then-using-daml-canton-console-to-access/4723

```
daml start
```

followed by

```
daml canton-console
```

doesn't work (because the console can't know which ports to connect to).

I'm very much not familiar at all with this area of the repo, so it's
possible this is not the right approach to solving that at all. Please
feel free to close and solve it in a better way if you know better.

CHANGELOG_BEGIN
- When using `daml start` (with no sandbox port configured in
  `daml.yaml`), the sandbox ports will no longer be random, and
  instead be the same ports as a plain `daml sandbox`.
CHANGELOG_END