Use rst_prolog to define typescript link targets
This allows us to define the typescript link targets as relative
or absolute depending on html/pdf output. This fixes the broken
links in the PDF output by making them absolute; the links
in the html output stay relative.

Reported by Ben on Slack: in the current PDF, the links are dead
(because they are relative links).

changelog_begin
changelog_end