docs: fix top-level update
There were two issues:

1. The `Directory.listFilesRecursive` function returns paths that
   include its first argument, so there was never any overlap between
   the sets.
2. That full path doesn't exist on AWS, and was given to AWS because
   `s3Path opts f` returns just `f` if it starts with `/`.

This means that, on the one hand, the script tried to delete
_everything_ from the top-level, which is a bit bad, but, on the other
hand, the script did not actually manage to delete anything at all,
which is basically the status quo.

Note that deleting everything isn't _that_ bad as we reupload right
after and we have a cache in front on the site, so this would only
impact users if they were the first to try and view a given page in that
small time window between deletion and reupload.

This PR should fix both issues by removing the version-specific prefix.

CHANGELOG_BEGIN
CHANGELOG_END