split up, parallelize assistant integration tests (#14281)
* parallelize assistant integration tests again by setting canton port options

Reduce the time of //daml-assistant/integration-tests:integration-tests
from 208 to 125 seconds locally by returning the test threading to 2,
replacing the setting to 1 in #12438 83f61bd19d75 with additional port
options to avoid conflicts.  The setting to 2 is explained in #9802.

* split quickstart from other assistant tests

Saves 20-30s (~15%) in total test runtime versus the parent.

* minimize hackage deps

CHANGELOG_BEGIN
CHANGELOG_END