release 2.3.2
This release is meant to include a fix on the Canton side. Because of
how our release process works at the moment, this means we have to
create a new release on this side too, even though there is no code
change.

So it is totally expected at this point that the 2.3.2 line points to
the same commit as the 2.3.1 line, even though it looks a bit strange.

CHANGELOG_BEGIN
CHANGELOG_END