daily: stop running JSON API perf tests
After discussion with the team, we've reached the conclusion that the
JSON API is not changing enough for these performance tests to bring
much value: we're not in a situation where we think there's a high risk
of performance regressions, nor trying to focus on performance
improvements.

Note that this is only removing the tests from the daily run; the team
does see value in keeping the code for these tests around in case any
performance-focused work does arise in the future.

(This PR also disables running the `check_releases` job outside of main
branch runs, because there's really never been a good reason to.)

Fixes #14608.

CHANGELOG_BEGIN
CHANGELOG_END

run-full-compat: true