Make emphasis visible (#14604)
Setting the font weight to 500 makes it effectively invisible, at least
in certain configurations (my setup is an Ubuntu 20.04 with Chrome 103).

In my experiments, 600 is the minimal font weight that makes the difference
visible on my setup. In order to keep the appearance of links unchanged,
I also decreased the font weight for links to the same as the body.

Another possibility to make a compromise that could work for different
setups could be to lower the font weight for bold to 600 and have a weight
specific for links at 500. This allows emphasis to appear clearly on setups
with less sophisticated font management but not to exagerated it for links.

changelog_begin
changelog_end