rs-grpc-bridge - Handle case where the response observer is cancelled before any elements are pushed through the gRPC layer.
If the producer for the BufferingSubscription never received elements then the stream was never marked as started.
Following downstream cancellation, the `BufferingSubscription.cancel` handle is called, but as the stream was never marked as running the cancellation is buffered for signaling in the DownstreamEventBuffer. As the DownstreamEventBuffer will flush the cancellation only if an element is received this leads to the stream remaining alive even if the observer was cancelled.

changelog_begin
changelog_end