Add "Next steps" page to "An introduction to Daml"
The idea is that of offering a comprehensive set of possible next steps after learning how to write Daml.

The page suggests certifications and points to existing examples as the previous final line on the last page,
as well as adding suggestions on where to learn more about building off-ledger services, how to install,
operate and maintain a ledger, where to learn about proven design patterns and the availability of Daml Hub
to get started without having to run your own ledger.

changelog_begin
changelog_end