move to daml subdir
We are planning to merge the daml repo into the (internal) canton repo.
As a preliminary step, both repos are moved into a subdirectory within
their current repo, which should make merging them (at the git level; no
plan to merge build systems or CI configuration at this stage) a lot easier.

This should be mostly transparent to the daml build as it finds its root
based on the `WORKSPACE` file, not the git root. The main issues are:

- Updating the paths to the YAML files in the Azure CI configuration
  (I've taken this opportunity to move `azure-cron.yml` and
  `azure-pipelines.yml` to the `ci` folder).
- Making sure everyone updates to the new structure locally. Changes to
  existing files should be brought over seamlessly through a simple
  rebase, but new files will need to be manually moved in their new place.

CHANGELOG_BEGIN
CHANGELOG_END