docs: no index
This removes the index.rst files from the repo, and instead generates a
placeholder index that just includes all the rst files at build time.

The reason for this change is that the ToC is defined in docs.daml.com,
and having one on both sides is both confusing and error-prone.

CHANGELOG_BEGIN
CHANGELOG_END