daily compat: fix BlackDuck invocation (#15004)
From output of currently failed jobs:

```
The key 'detect.npm.dependency.types.excluded' in property source
'commandLineArgs' contained a value that could not be reasonably
converted to the properties type. The exception was: Unable to parse raw
value 'NON_PRODUCTION' and coerce it into type 'NoneEnum or
NpmDependencyType'. Value was must be one of NONE,DEV,PEER
```

It looks like I was a bit overeager on #14995.

CHANGELOG_BEGIN
CHANGELOG_END