review daml-intro-10 (#15268)
I started with wanting to remove the warning on incomplete pattern
match, thinking example code should be "good". Then I decided I should
skim the corresponding docs to check if that warning, or the pragma,
were actually mentioned in the text.

It turns out the warning should be there, but the pragma is not
explained and not necessary to get the warning. So the pargma can go.

But in the process I noticed a couple things that I belive can be
improved in the wording of the document, so I went agead and made a few
changes.

CHANGELOG_BEGIN
CHANGELOG_END