do not use the external definition of "serializable"
"ones whose values can be written in a text or binary format" is a very
misleading definition to give for "serializable types" in the context of
Daml.

When we say "serializable" in Daml, we *always* mean the formal Daml-LF
definition given in "Serializable types" in daml-lf-1.rst, i.e. ⊢ₛ τ, or
"τ derives from the rules S".  That is also what we mean here.

There are many examples of types "whose values can be written in a text
or binary format", that are nevertheless *not* derivable from S.

As it says in daml-lf-1.rst, "serializable types are the types whose
values can be persisted on the ledger".  So we copy over something like
that language.

CHANGELOG_BEGIN
CHANGELOG_END