Remove outdated docs referring to 1.18.x features
The development of the Helm Chart halted after 1.18.x to make room
for other priorities, including a better documentation of operational
concerns as a way for users to bridge the gap we were originally
trying to fill with the Helm Chart. The pages remained in the
documentation as we were hoping for the feature to go back into
active development but since the feature is not actually working and
we do not have short term plans to revive it (and possibly not in
the current form) it is better to just yank the content until
we have a better understanding of where we want this to go.

If by any chance we want to restore the feature more or less in
its current form, restoring this document from a previous version
requires very little effort while simultaneously we do not cause
confusion by documenting a feature that does not work with the
current version.

changelog_begin
changelog_end