Drop comments about 0.x when talking about upgrades
The comments refer to 0.x and can be misleading for 2.x users
(since the default is not 1.8 anymore and have not been for a
long time at this point). Furthermore, I am also relatively
sure that 1.8 is not even an option at all for 2.x users.

changelog_begin
changelog_end