[BACKPORT #15311] dev-env: pin msys2/mingw (#15491)
Today the postgresql 12.4 artifacts were purged from the msys2 primary
mirrors. I was able to find a copy on some old mirrors (I downloaded
from [mirrors.piconets.webwerks.in], [mirrors.aliyun.com], and
[mirror.iscas.ac.cn]), but that raised the question of where to put the
binary so our CI nodes can access it. I decided to put it on one of our
GCS buckets, and while I was there I did the whole msys2 set of
packages.

Does this manually may not be the best path forward for a long-term
solution (though maybe it is; they tend not to change all that often),
but it does unblock us right now.

For the postgresql binary specifically, I have checked that the three
mirrors give me the exact same file between them, and that it matches
the old hash. For the other files, I downloaded them from the primary
mirror ([repo.msys2.org]) and checked that the hashes match the recorded
values in our JSON manifest.

CHANGELOG_BEGIN
CHANGELOG_END

[mirrors.piconets.webwerks.in]: https://mirrors.piconets.webwerks.in/msys2-mirror/mingw/x86_64/mingw-w64-x86_64-postgresql-12.4-1-any.pkg.tar.zst
[mirrors.aliyun.com]: https://mirrors.aliyun.com/msys2/mingw/x86_64/mingw-w64-x86_64-postgresql-12.4-1-any.pkg.tar.zst?spm=a2c6h.25603864.0.0.32b23240Qp8jCL
[mirror.iscas.ac.cn]: https://mirror.iscas.ac.cn/msys2/mingw/x86_64/mingw-w64-x86_64-postgresql-12.4-1-any.pkg.tar.zst
[repo.msys2.org]: https://repo.msys2.org/mingw/x86_64/