release 2.4.2 (#15515)
Note: this _replaces_ 2.4.1 because, _if_ we want to have #15511 in the
next release, it doesn't really make sense (in my opinion) to go through
the rest of the 2.4.1 process.

Happy to discuss, though: it's possible we still want both releases,
it's also possible we don't really need #15511. I really don't have
enough context to say, so I'm just opening this to provide the option.