Fix error handling in time service
updateTime can return an error and the code correctly tried to turn
that into invalidArgument and logged it. However, the Either returned
by that then got completely discarded and the overall request
succeeded which definitely seems like a bug.

I added a test and verified that it fails withotu this change.

The code seems fairly messy. There are a few reasons for that:

1. We use Future[Either[…]] and we treat failures in the Either
   differently from future failures. I'm not quite convinced that
   makes sense but for this PR I didn’t want to introduce unrelated
   changed.
2. Future[Either[…]] without some form of EitherT is inherently messy
   but afaik you don’t use cats here so I kept it like this.