Fix flaky gRPC server metrics test
The assertions are wrapped in an `eventually` because we first call the `delegate` methods (eg: using `super.close`) and only afterwards we record the metrics.
This means that metric recording does not impact the call before we respond, but it also means that there's a tiny delay between returning the answer and the metrics actually getting updated.