Only warn about errors loading test suites by default
It turned out that throwing an exception is too strict for the downstream Canton project
which currently fails on CI caused by errors in a number of test suites.

Until they have adapted (see https://github.com/DACH-NY/canton/issues/11083) we only warn about
the errors by default.

For this project we ensure that warnings are fatal.