temporary workaround for stream shutdown, gets a little farther
S11 trying to abort The stream has not been completed in 2 seconds.
S11 finished in wsMessageHandler The stream has not been completed in 2 seconds.
S11 trying to abort The stream has not been completed in 2 seconds.
S11 trying to abort The stream has not been completed in 2 seconds.
S11 trying to abort The stream has not been completed in 2 seconds.
S11 trying to abort The stream has not been completed in 2 seconds.
S11 trying to abort The stream has not been completed in 2 seconds.
S11 trying to abort The stream has not been completed in 2 seconds.
S11 trying to abort The stream has not been completed in 2 seconds.
S11 finished in wsMessageHandler The stream has not been completed in 2 seconds.
S11 finished in wsMessageHandler The stream has not been completed in 2 seconds.
S11 finished in wsMessageHandler The stream has not been completed in 2 seconds.
S11 finished in wsMessageHandler The stream has not been completed in 2 seconds.
S11 finished in wsMessageHandler The stream has not been completed in 2 seconds.