ignore build-and-lint-test `yarn.lock` inconsistency on release
With the below command, with or without the entry for
@types/prettier@2.6.0 removed, this command passes.

DAML_SDK_RELEASE_VERSION=0.0.0-head bazel test --test_output=streamed //language-support/ts/codegen/tests:build-and-lint-test

With the below command with the entry for @types/prettier@2.6.0 removed,
this command fails with the expected error about a missing entry.  With
the entry restored, it passes.

bazel test --test_output=streamed //language-support/ts/codegen/tests:build-and-lint-test