cron: fix sgnature key
I _believe_ this is what has been tripping up the check-releses job for
the past few days: while updating the signing key, I did not update the
key fingerprint to match, and therefore trying to check any signature
fails.

The failure mode of the existing script is still pretty bad, regadless:
looping on the same release until the hard drive is full is definitely
not expected. Our retry policies are supposed to have time bounds, and
failure should result in cleaning up the workdir.

run-full-compat: true