[DPP-1400] Refactor and simplify the pruning backend tests (#16220)
A concern with the existing pruning backend tests is that they are relatively complex particularly because:
- in order to fetch data from the indexdb they use production sql queries which might have builtin knowledge of pruning offsets,
- production sql queries are not designed for ergonomic one-off usage.

This PR introduces a set of queries, one for each table subject to pruning, each returning a record for each row its table. This leads to more a concise and straightforward way assert what data has or hasn't been pruned.