infra: larger virtual disks
I've seen a few "disk full" errors recently on CI, and they don't quite
make sense because we do have this "reset cache" step that looks at
available disk space and cleans up if needed.

So I dug a little bit more and found this discrepancy: the real hard
drives are 400g, the virtual disks are 200g. So what may happen is we
still have plenty of space on the real drives (well, they're virtual,
but bear with me), while the virtual ones are full. And since the
clean-up step only looks at the free space on the real drives, it goes
ahead without cleaning up and then when we try to download stuff there's
no free space on the virtual drives.

This didn't use to be a problem because the size of the virtual drives
mapped to the size of the real ones. This PR aims at restoring that
equivalence.