fix blackduck
BlackDuck is currently failing to run because it creates multiple report
files and then expects a single one. This is due to #16552 changing only
one of the output file names.

This PR resolves this by:

1. Changing the name everywhere, so all runs aggregate their results on
   the same file.
2. Changing the PR step code to expect the given file name, rather than
   try and glob it.

My understanding of the BlackDuck setup is very fuzzy so please
double-check carefully.

See #16600 for an example of what the update would be from this PR.