[backport #16683] support numerous observers in JSON/Oracle (#16707)
When the `observers` list was too long, we were getting an `ORA-40478: output value too large (maximum 4000)` error from Oracle.

This was being hit when the `contract_stakeholders` view was being updated. An intermediate part of the expression which updates the view was being defaulted to a type which didn't allow larger values. Adding an explicit type fixes the problem.

- Added test which could reproduce the previous failure.
- Moved `randomTextN` to be accessible to the new test.

Fixes https://digitalasset.atlassian.net/browse/LT-11

Co-authored-by: Raphael Speyer <raphael.speyer@digitalasset.com>