JSON API: fix empty string comparisons on Oracle (#16892)
Oracle conflates empty strings with NULL, which breaks comparison operations against empty strings.

This change makes Oracle-backed queries have the behaviour you'd expect when comparing empty strings, in line with what we see on Postgres and in-memory backed queries.

We had to take a bit of care to ensure it worked irrespective of whether the JSON index was enabled.

Fixes https://digitalasset.atlassian.net/browse/LT-24