Simplify Anf.scala by removing the continuations.
The core functions of the ANF transformation in Anf.scala handle 3
different continuations:
  1. The "transformations" (called tx), that are an inherent aspect of
     any ANF transformation expressed in pseudo-CPS style for
     convenience.
  2. The "continuations" (called k) that had been introduced to avoid
     stack oveflow.
  3. The trampolines, that have also been introduced to avoid stack
     overflow.

This commit recognize that types 2 and 3 are redundant and merges them
into only type 3. It relies on the fact that trampolines are monadic,
and so we switch from the hand-rolled trampolines to that of scala std
library since they already implement flatMap.