Perform full ANF in Anf.scala whenever enableFullAnfTransformation is true.
There are 11 mutually recursive methods (transitively) affected by that flag
so I've turned the object into a class. Unfortunately, classes that extend
AnyVal cannot be members of another class so I had to move the to the object,
which makes for a weird artificial split. To make it less artificial, I've
moved to the object all methods that are static. It is somewhat
unsatisfactory to split the code like that but fortunately all the methods
that go into the objects are helper methods and all the ones that remain
in the class constitute the core of the transformation.