Test both evaluation orders in daml script tests. (#17271)
PR #17253 introduces a right-to-left evaluation order behind an interpreter feature flag. It also modifies every scala test that uses the interpreter to test both evaluation orders and documents any divergence as a new test case.

In this commit we do the same for Daml script tests (compiler/damlc/tests/daml-test-files/...). We introduce a new file annotation for Daml script tests: @EVALUATION-ORDER (LeftToRight|RightToLeft). A test that uses this annotation will be ignored when run with an evaluation order that differs from the specified one. The BUILD file is modified to run every test twice: once in left-to-right order, and once in right-to-left order.

The only test for which the evaluation orders disagree is SemanticsEvalOrder.daml. So we mark it as left-to-right only, and we introduce SemanticsEvalOrderRightToLeft.daml, which is right-to-left only and documents the new behavior. All the other existing tests are run using both modes and do not diverge.