release.sh: add rc, remove new (#17280)
The goal of this change is to make it possible to do explicit "rc"
releases from the daml repo. We've had a good run at the old way of
working, where RCs were supposed to be "whatever the latest snapshot is
at the time of release", but since we've moved away from time-based
releases that attitude has become less and less in line with reality.

And if we're deliberately creating RCs now, why not call them that?