[Speedy] Ensure evaluation of Update questions do not throw
It is important the continuations of Engine Result does not throw
exceptions, as we cannot ensure those are caught properly.

Until now, we were carefully taking care the code we write can never
throw exceptions. However as the code in continuation grows, I think
this is not possible. For this reason we deveop a systemic approch and
caught any possible exception throw in the continuation, and delay the
error reporting to the next Speedy iteration.