[Speedy] Ensure continuation of Update questions do not throw (#17549)
It is important continuations of Engine Result does not throw exceptions, as we cannot ensure those are properly caught.

Until now, we have to carefully review the code we write never throws exceptions. However as the code in continuation grows, I think this is not possible to do it anymore. For this reason, this PR systematically caught any possible exception thrown during continuation evaluation, and delay the error reporting to the next Speedy iteration.


Co-authored-by: Moisés Ackerman <6054733+akrmn@users.noreply.github.com>