Make `TransactionFilter.toProto` public (#17809)
* Make `TransactionFilter.toProto` public

We are in the process of enabling the usage of the Java bindings from the
Canton Console. In order to do so in a backward-compatible fashion, we
often rely on existing Scala code and use the Protobuf representation as
a bridge between the two. Unfortunately the fact that the `toProto` method
on `TransactionFilter`s is package-private forced us to temporarily add
a conversion method located within the package that re-exported it publicly
for usage as part of the Canton Console. This change should enable us to
remove this workaround. I would be happy to hear other possible approaches
to consider.

* Update subclass