Add example of sslcontext based on a keystore, which does work.
The issue is that currently we do our config based on the certificate, private key, and ca file seperately, rather than via a keystore + password.

Perhaps for now we could change the HTTPS config to be based on keystore + password, and then figure out how to support the other mode in due course.