JSON API: add new PrunedOffset error for websocket queries that attempt to use an offset that has been pruned (#17905)
This means that if a websocket query is initiated with an explicit offset, and the ledger returns an error reporting that the offset has been pruned, the websocket query will be terminated with an error containing a status of `410` (Gone) and a message indicating that the offset has been pruned.

The client will need to react to this by clearing any state that had been incrementally built based on updates, and refresh with current version of ledger state.

This is related to https://github.com/digital-asset/daml/issues/13788 and https://github.com/digital-asset/daml/issues/13680

As discussed there, the existing akka APIs do not support setting the websocket close code and reason. The approach taken here is to return a single message with a custom error and close the websocket.