JSON API HTTPS: use a library to be more flexible in the private keys we support (#18047)
The original implementation parsed the private key file manually to avoid adding any library deps in the middle of the CVE-quake, where libraries were already being changed. To avoid inscrutable errors, I added an assertion to ensure the input was exactly what was expected.
Now that dependency flux has settled down, we can switch to BouncyCastle, which is more flexible in its formats, and remove the extra assertion.